Refactor logic's config
