+++ packages/sanes-chrome-extension/src/logic/blockchain/chainsConfig/config.test.ts
